% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{maorinames}
\alias{maorinames}
\title{Maori names.}
\format{A data frame with 6 variables: \code{Name}, \code{Sex},
\code{Source}, \code{Year}, \code{Born} and \code{Rank}}
\usage{
maorinames
}
\description{
The  top 20 Māori baby names for 2016 and 2015 by sex.
In 2015 data only includes children born in NZ, in 2016 rankings Māori baby
names among children born overseas are also included
}
\keyword{datasets}
