% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nzbabynames-package.R
\docType{package}
\name{nzbabynames-package}
\alias{nzbabynames}
\alias{nzbabynames-package}
\title{nzbabynames: What the Package Does (One Line, Title Case)}
\description{
Data on the most popular baby names in New Zealand.
}
\author{
\strong{Maintainer}: Emily Kothe \email{emily.kothe@deakin.edu.au}

}
\keyword{internal}
