% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{nzbabynames}
\alias{nzbabynames}
\title{Baby names.}
\format{A data frame with five variables: \code{Year}, \code{Sex},
\code{Name}, \code{Count}.}
\usage{
nzbabynames
}
\description{
Full baby name data NZ. This includes all names with at
least 10 uses.
}
\keyword{datasets}
