% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/babynames.R
\docType{data}
\name{nzbabynames}
\alias{nzbabynames}
\title{Top 100 baby names annually in New Zealand 1954-2016}
\format{A data frame with four variables: ‘year’, ‘sex’, ‘name’ and ‘n’.}
\source{
\url{https://smartstart.services.govt.nz/news/baby-names/}
}
\usage{
nzbabynames
}
\description{
A dataset of the top 100 baby names annually in New Zealand since 1954,
provided by the Department of Internal Affairs.
}
\keyword{datasets}
