% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{nzbirths}
\alias{nzbirths}
\title{nzbirths}
\format{An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 84 rows and 4 columns.}
\usage{
nzbirths
}
\description{
Data on live births from 1935-2017. Data from 1971 includes live births by
sex. \emph{Note: Data before 1991 includes New Zealand residents and overseas
visitors. Data from 1991 includes New Zealand residents only.}
}
\keyword{datasets}
