% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/births.R
\docType{data}
\name{nzbirths}
\alias{nzbirths}
\title{Annual births in New Zealand 1954-2016}
\format{A data frame with two variables: ‘year’ and ‘births’.}
\source{
Statistics New Zealand
}
\usage{
nzbirths
}
\description{
Live births data from Statistics New Zealand.  Includes births registered
overseas.
}
\keyword{datasets}
